+++ lib/action_client/middleware/tagger.rb
@@ -1,7 +1,0 @@
-module ActionClient
-  module Middleware
-    Tagger = proc do |request|
-      "ActionClient - #{request.request_method} - #{request.original_url}"
-    end
-  end
-end